# perf(es/parser): Improve performance by using `#[cold]`