--- a/crates/swc_ecma_parser/src/parser/macros.rs
+++ b/crates/swc_ecma_parser/src/parser/macros.rs
@@ -247,14 +247,13 @@ macro_rules! cur {
             Some(c) => Ok(c),
             None => {
                 if $required {
-                    let err = crate::error::Error {
-                        error: Box::new((last, crate::error::SyntaxError::Eof)),
-                    };
+                    let err = crate::error::Error::new(last, crate::error::SyntaxError::Eof);
                     return Err(err);
                 }
-                Err(crate::error::Error {
-                    error: Box::new((last, crate::error::SyntaxError::Eof)),
-                })
+                Err(crate::error::Error::new(
+                    last,
+                    crate::error::SyntaxError::Eof,
+                ))
             }
         }
     }};
@@ -274,9 +273,7 @@ Current token is {:?}",
         match $p.input.peek() {
             Some(c) => Ok(c),
             None => {
-                let err = crate::error::Error {
-                    error: Box::new((last, crate::error::SyntaxError::Eof)),
-                };
+                let err = crate::error::Error::new(last, crate::error::SyntaxError::Eof);
                 Err(err)
             }
         }