--- a/crates/swc_ecma_parser/src/parser/tests.rs
+++ b/crates/swc_ecma_parser/src/parser/tests.rs
@@ -341,16 +341,14 @@ fn illegal_language_mode_directive2() {
             let errors = p.take_errors();
             assert_eq!(
                 errors,
-                vec![Error {
-                    error: Box::new((
-                        Span {
-                            lo: BytePos(21),
-                            hi: BytePos(34),
-                            ctxt: swc_common::SyntaxContext::empty()
-                        },
-                        crate::parser::SyntaxError::IllegalLanguageModeDirective
-                    ))
-                }]
+                vec![Error::new(
+                    Span {
+                        lo: BytePos(21),
+                        hi: BytePos(34),
+                        ctxt: swc_common::SyntaxContext::empty()
+                    },
+                    crate::parser::SyntaxError::IllegalLanguageModeDirective
+                )]
             );
 
             Ok(program)