--- a/crates/swc_ecma_parser/src/parser/tests.rs
+++ b/crates/swc_ecma_parser/src/parser/tests.rs
@@ -314,16 +314,14 @@ fn illegal_language_mode_directive1() {
             let errors = p.take_errors();
             assert_eq!(
                 errors,
-                vec![Error {
-                    error: Box::new((
-                        Span {
-                            lo: BytePos(20),
-                            hi: BytePos(33),
-                            ctxt: swc_common::SyntaxContext::empty()
-                        },
-                        crate::parser::SyntaxError::IllegalLanguageModeDirective
-                    ))
-                }]
+                vec![Error::new(
+                    Span {
+                        lo: BytePos(20),
+                        hi: BytePos(33),
+                        ctxt: swc_common::SyntaxContext::empty()
+                    },
+                    crate::parser::SyntaxError::IllegalLanguageModeDirective
+                )]
             );
 
             Ok(program)