**Description:**

We emulate `unlikely` for vairous error macros by converting `Error::new` into a cold function.

## Before

```
test angular       ... bench:   7,922,906 ns/iter (+/- 284,640) = 90 MB/s
test backbone      ... bench:   1,093,366 ns/iter (+/- 30,800) = 54 MB/s
test colors        ... bench:      14,564 ns/iter (+/- 635) = 79 MB/s
test jquery        ... bench:   6,021,029 ns/iter (+/- 209,889) = 44 MB/s
test jquery_mobile ... bench:   9,654,149 ns/iter (+/- 154,489) = 46 MB/s
test mootools      ... bench:   4,645,033 ns/iter (+/- 117,356) = 34 MB/s
test three         ... bench:  27,476,366 ns/iter (+/- 1,396,719) = 43 MB/s
test underscore    ... bench:     942,445 ns/iter (+/- 12,220) = 46 MB/s
test yui           ... bench:   4,686,897 ns/iter (+/- 180,492) = 72 MB/s
```



## After

```
test angular       ... bench:   7,823,925 ns/iter (+/- 214,984) = 91 MB/s
test backbone      ... bench:   1,077,099 ns/iter (+/- 24,427) = 55 MB/s
test colors        ... bench:      14,139 ns/iter (+/- 88) = 81 MB/s
test jquery        ... bench:   5,954,287 ns/iter (+/- 167,003) = 45 MB/s
test jquery_mobile ... bench:   9,538,154 ns/iter (+/- 266,640) = 47 MB/s
test mootools      ... bench:   4,581,866 ns/iter (+/- 112,853) = 35 MB/s
test three         ... bench:  27,164,424 ns/iter (+/- 491,544) = 43 MB/s
test underscore    ... bench:     927,293 ns/iter (+/- 14,810) = 46 MB/s
test yui           ... bench:   4,626,075 ns/iter (+/- 141,111) = 73 MB/s
```